requirements.txt
anyio==3.6.2
asyncpg==0.27.0
black==23.3.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
ecdsa==0.18.0
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy==1.2.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-dotenv==0.5.2
python-dotenv==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.14
sqlparse==0.4.4
starlette==0.26.1
tabulate==0.9.0
typing_extensions==4.5.0
uvicorn==0.21.1
yoyo-migrations==8.2.0
zipp==3.15.0
